# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Páginas Abiertas book club portal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type VotingSessionStatus =
  (typeof VotingSessionStatus)[keyof typeof VotingSessionStatus];

export const VotingSessionStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface VotingSession {
  id: number;
  title: string;
  status: VotingSessionStatus;
  deadline?: string | null;
  createdAt: string;
}

export interface VotingSessionResponse {
  session: VotingSession | null;
}

export interface CandidateBook {
  id: number;
  sessionId: number;
  title: string;
  author: string;
  genre: string;
  coverUrl: string;
  synopsis: string;
  votes: number;
  isWinner: boolean;
}

export interface CastVoteBody {
  code: string;
  bookId: number;
  secondBookId?: number | null;
}

export interface VoteResult {
  success: boolean;
  message: string;
  votesUsed: number;
}

export type VotingCodeType =
  (typeof VotingCodeType)[keyof typeof VotingCodeType];

export const VotingCodeType = {
  standard: "standard",
  premium: "premium",
} as const;

export interface VotingCode {
  id: number;
  sessionId: number;
  code: string;
  type: VotingCodeType;
  used: boolean;
  usedAt?: string | null;
}

export type LeaderboardMemberRank =
  (typeof LeaderboardMemberRank)[keyof typeof LeaderboardMemberRank];

export const LeaderboardMemberRank = {
  Novata: "Novata",
  Viajera: "Viajera",
  Aventurera: "Aventurera",
  Embajadora: "Embajadora",
  Leyenda_Literaria: "Leyenda Literaria",
} as const;

export interface LeaderboardMember {
  id: number;
  alias: string;
  avatar: string;
  points: number;
  rank: LeaderboardMemberRank;
  position: number;
  archived: boolean;
}

export interface CurrentBook {
  id: number;
  title: string;
  author: string;
  genre: string;
  coverUrl: string;
  synopsis: string;
  currentWeek: number;
  totalWeeks: number;
  nextSessionDate: string;
  nextSessionDescription: string;
  weekActivity: string;
  motivationalPhrase: string;
}

export interface CurrentBookResponse {
  book: CurrentBook | null;
}

export interface CreateVotingSessionBody {
  title: string;
  deadline?: string | null;
}

export interface AddBookBody {
  title: string;
  author: string;
  genre: string;
  coverUrl: string;
  synopsis: string;
}

export type GenerateCodesBodyType =
  (typeof GenerateCodesBodyType)[keyof typeof GenerateCodesBodyType];

export const GenerateCodesBodyType = {
  standard: "standard",
  premium: "premium",
} as const;

export interface GenerateCodesBody {
  quantity: number;
  type: GenerateCodesBodyType;
}

export interface AddMemberBody {
  alias: string;
  avatar: string;
  points: number;
}

export interface UpdateMemberBody {
  alias?: string;
  avatar?: string;
  points?: number;
}

export interface AdminVerifyBody {
  password: string;
}

export interface AdminVerifyResponse {
  success: boolean;
}

export interface UpdateCurrentBookBody {
  title: string;
  author: string;
  genre: string;
  coverUrl: string;
  synopsis: string;
  currentWeek: number;
  totalWeeks: number;
  nextSessionDate: string;
  nextSessionDescription: string;
  weekActivity: string;
  motivationalPhrase: string;
}

export interface LiteraryCountry {
  id: number;
  name: string;
  emoji: string;
  description: string;
  color: string;
  booksRead: number;
  displayOrder: number;
  updatedAt: string;
}

export interface UpdateLiteraryCountryBody {
  name?: string;
  emoji?: string;
  description?: string;
  color?: string;
  booksRead?: number;
}