# lib/api-zod/src/generated/types/literaryCountry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Páginas Abiertas book club portal
 * OpenAPI spec version: 0.1.0
 */

export interface LiteraryCountry {
  id: number;
  name: string;
  emoji: string;
  description: string;
  color: string;
  booksRead: number;
  displayOrder: number;
  updatedAt: string;
}